+++ google/cloud/storage/grpc_plugin_test.cc
@@ -21,7 +21,7 @@
 
 namespace google {
 namespace cloud {
-namespace storage_experimental {
+namespace storage {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 namespace {
 
@@ -45,8 +45,7 @@ TEST(GrpcPluginTest, MostConfigValuesCreatesGrpc) {
       ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", absl::nullopt);
   // Unless the config is set to "none" we want to create the gRPC stub.
   for (auto const* config : {"", "metadata", "media", "anything-but-none"}) {
-    auto client =
-        DefaultGrpcClient(TestOptions().set<GrpcPluginOption>(config));
+    auto client = MakeGrpcClient(TestOptions().set<GrpcPluginOption>(config));
     auto impl = ClientImplDetails::GetConnection(client);
     ASSERT_THAT(impl, NotNull());
     EXPECT_THAT(impl->InspectStackStructure(),
@@ -60,8 +59,7 @@ TEST(GrpcPluginTest, EnvironmentOverrides) {
       ScopedEnvironment("CLOUD_STORAGE_ENABLE_TRACING", absl::nullopt);
   auto config =
       ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", "none");
-  auto client =
-      DefaultGrpcClient(TestOptions().set<GrpcPluginOption>("metadata"));
+  auto client = MakeGrpcClient(TestOptions().set<GrpcPluginOption>("metadata"));
   auto impl = ClientImplDetails::GetConnection(client);
   ASSERT_THAT(impl, NotNull());
   EXPECT_THAT(impl->InspectStackStructure(),
@@ -74,7 +72,7 @@ TEST(GrpcPluginTest, UnsetConfigCreatesMetadata) {
       ScopedEnvironment("CLOUD_STORAGE_ENABLE_TRACING", absl::nullopt);
   auto config =
       ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", absl::nullopt);
-  auto client = DefaultGrpcClient(TestOptions());
+  auto client = MakeGrpcClient(TestOptions());
   auto impl = ClientImplDetails::GetConnection(client);
   ASSERT_THAT(impl, NotNull());
   EXPECT_THAT(impl->InspectStackStructure(),
@@ -87,7 +85,7 @@ TEST(GrpcPluginTest, NoneConfigCreatesCurl) {
       ScopedEnvironment("CLOUD_STORAGE_ENABLE_TRACING", absl::nullopt);
   auto config =
       ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", absl::nullopt);
-  auto client = DefaultGrpcClient(TestOptions().set<GrpcPluginOption>("none"));
+  auto client = MakeGrpcClient(TestOptions().set<GrpcPluginOption>("none"));
   auto impl = ClientImplDetails::GetConnection(client);
   ASSERT_THAT(impl, NotNull());
   EXPECT_THAT(impl->InspectStackStructure(),
@@ -101,15 +99,30 @@ TEST(GrpcPluginTest, HybridUsesGrpcBufferOptions) {
       ScopedEnvironment("CLOUD_STORAGE_ENABLE_TRACING", absl::nullopt);
   auto config =
       ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", absl::nullopt);
-  auto client = DefaultGrpcClient(TestOptions().set<GrpcPluginOption>("media"));
+  auto client = MakeGrpcClient(TestOptions().set<GrpcPluginOption>("media"));
   EXPECT_GE(
       client.raw_client()->options().get<storage::UploadBufferSizeOption>(),
       32 * 1024 * 1024L);
+}
+
+TEST(GrpcPluginTest, BackwardsCompatibilityShims) {
+  // Explicitly disable logging, which may be enabled by our CI builds.
+  auto logging =
+      ScopedEnvironment("CLOUD_STORAGE_ENABLE_TRACING", absl::nullopt);
+  auto config =
+      ScopedEnvironment("GOOGLE_CLOUD_CPP_STORAGE_GRPC_CONFIG", absl::nullopt);
+  namespace gcs_ex = ::google::cloud::storage_experimental;
+  auto client = gcs_ex::DefaultGrpcClient(
+      TestOptions().set<gcs_ex::GrpcPluginOption>("none"));
+  auto impl = ClientImplDetails::GetConnection(client);
+  ASSERT_THAT(impl, NotNull());
+  EXPECT_THAT(impl->InspectStackStructure(),
+              ElementsAre("RestStub", "StorageConnectionImpl"));
 }
 #include "google/cloud/internal/diagnostics_pop.inc"
 
 }  // namespace
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
-}  // namespace storage_experimental
+}  // namespace storage
 }  // namespace cloud
 }  // namespace google